requirements.txt
appdirs==1.4.2
Django==1.10.5
django-allauth==0.30.0
django-haystack==2.6.0
httplib2==0.10.3
oauth2==1.9.0.post1
oauthlib==2.0.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
PyJWT==1.4.2
pyparsing==2.1.10
python-openid==2.2.5
python-social-auth==0.2.21
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
social-auth-core==1.2.0
